Cleanups and README updates.
