+++ src/main/java/org/mozilla/iot/webthing/Property.java
@@ -1,7 +1,6 @@
 /**
  * High-level Property base class implementation.
  */
-
 package org.mozilla.iot.webthing;
 
 import org.json.JSONObject;
@@ -27,9 +26,10 @@ public class Property<T> {
      *
      * @param thing Thing this property belongs to
      * @param name  Name of the property
+     * @param value Value object to hold the property value
      */
-    public Property(Thing thing, String name) {
-        this(thing, name, null, null);
+    public Property(Thing thing, String name, Value<T> value) {
+        this(thing, name, value, null);
     }
 
     /**
@@ -37,17 +37,14 @@ public class Property<T> {
      *
      * @param thing    Thing this property belongs to
      * @param name     Name of the property
+     * @param value    Value object to hold the property value
      * @param metadata Property metadata, i.e. type, description, unit, etc., as
      *                 a Map
      */
-    public Property(Thing thing, String name, Map<String, Object> metadata) {
-        this(thing, name, metadata, null);
-    }
-
     public Property(Thing thing,
                     String name,
-                    Map<String, Object> metadata,
-        Value<T> value) {
+                    Value<T> value,
+                    Map<String, Object> metadata) {
         this.thing = thing;
         this.name = name;
         this.value = value;
@@ -59,8 +56,10 @@ public class Property<T> {
         } else {
             this.metadata = metadata;
         }
-        //Adds the property change observer, to notify the Thing about a property change
-        this.value.addObserver((a,b)->this.thing.propertyNotify(this));
+
+        // Add the property change observer to notify the Thing about a
+        // property change
+        this.value.addObserver((a, b) -> this.thing.propertyNotify(this));
     }
 
     /**
