+++ src/main/java/org/mozilla/iot/webthing/Thing.java
@@ -1,7 +1,6 @@
 /**
  * High-level Thing base class implementation.
  */
-
 package org.mozilla.iot.webthing;
 
 import org.json.JSONArray;
