+++ src/main/java/org/mozilla/iot/webthing/Utils.java
@@ -1,7 +1,6 @@
 /**
  * Utility functions.
  */
-
 package org.mozilla.iot.webthing;
 
 import java.net.Inet4Address;
