+++ src/test/java/org/mozilla/iot/webthing/example/SingleThingTest.java
@@ -2,10 +2,7 @@ package org.mozilla.iot.webthing.example;
 
 import org.junit.Test;
 
-/**
- * @author Tim Hinkes (timmeey@timmeey.de)
- */
-public class TestServerExample {
+public class SingleThingTest {
     @Test
     public void testTestServer() {
     }
